@@ -15,18 +15,18 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 3.0.0+1
+version: 3.1.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
   flutter_rating_bar: ^4.0.0
-  flutter_swiper: ^1.1.6
-  google_fonts: ^2.0.0
+  card_swiper: ^1.0.2
+  google_fonts: ^2.1.0
   shimmer: ^2.0.0
 
 dev_dependencies:
