@@ -1,10 +1,8 @@
-import 'package:flutter/foundation.dart';
-
 class AvailableCoupon {
   const AvailableCoupon({
-    @required this.coupon,
-    @required this.discount,
-    @required this.desc,
+    required this.coupon,
+    required this.discount,
+    required this.desc,
   });
 
   final String coupon;
@@ -16,37 +14,44 @@ class AvailableCoupon {
       AvailableCoupon(
         coupon: '100INDUSIND',
         discount: 'Get 20 % discount using Induslnd Bank Credit Cards',
-        desc: 'Use code 100INDUSIND & get 20 % discount up to Rs100 on orders above Rs400',
+        desc:
+            'Use code 100INDUSIND & get 20 % discount up to Rs100 on orders above Rs400',
       ),
       AvailableCoupon(
         coupon: 'HSBC500',
         discount: 'Get 15 % discount using HSBC Bank Credit Cards',
-        desc: 'Use code HSBC500 & get 14 % discount up to Rs125 on orders above Rs500',
+        desc:
+            'Use code HSBC500 & get 14 % discount up to Rs125 on orders above Rs500',
       ),
       AvailableCoupon(
         coupon: '100TMB',
         discount: 'Get 20 % discount using TMB Bank Credit Cards',
-        desc: 'Use code 100TMB & get 20 % discount up to Rs100 on orders above Rs400',
+        desc:
+            'Use code 100TMB & get 20 % discount up to Rs100 on orders above Rs400',
       ),
       AvailableCoupon(
         coupon: 'INDUSIND20',
         discount: 'Get 20 % discount using Induslnd Bank Credit Cards',
-        desc: 'Use code INDUSIND20 & get 20 % discount up to Rs200 on orders above Rs600',
+        desc:
+            'Use code INDUSIND20 & get 20 % discount up to Rs200 on orders above Rs600',
       ),
       AvailableCoupon(
         coupon: 'HSBC5100',
         discount: 'Get 20 % discount using Induslnd Bank Credit Cards',
-        desc: 'Use code HSBC5100 & get 20 % discount up to Rs100 on orders above Rs400',
+        desc:
+            'Use code HSBC5100 & get 20 % discount up to Rs100 on orders above Rs400',
       ),
       AvailableCoupon(
         coupon: '100AXIS',
         discount: 'Get 20 % discount using Axis Bank Credit Cards',
-        desc: 'Use code 100AXIS & get 20 % discount up to Rs100 on orders above Rs400',
+        desc:
+            'Use code 100AXIS & get 20 % discount up to Rs100 on orders above Rs400',
       ),
       AvailableCoupon(
         coupon: '100INDUSIND',
         discount: 'Get 20 % discount using Induslnd Bank Credit Cards',
-        desc: 'Use code 100INDUSIND & get 20 % discount up to Rs100 on orders above Rs400',
+        desc:
+            'Use code 100INDUSIND & get 20 % discount up to Rs100 on orders above Rs400',
       ),
     ];
   }
