@@ -1,9 +1,7 @@
-import 'package:flutter/foundation.dart';
-
 class IndianFood {
   const IndianFood({
-    @required this.image,
-    @required this.name,
+    required this.image,
+    required this.name,
   });
 
   final String image;
