@@ -1,9 +1,7 @@
-import 'package:flutter/foundation.dart';
-
 class RestaurantDetail {
   const RestaurantDetail({
-    @required this.title,
-    @required this.price,
+    required this.title,
+    required this.price,
     this.image = '',
     this.desc = '',
   });
@@ -19,7 +17,8 @@ class RestaurantDetail {
         title: 'Idly(2Pcs) (Breakfast)',
         price: 'Rs48',
         image: 'assets/images/food1.jpg',
-        desc: 'A healthy breakfast item and an authentic south indian delicacy! Steamed and fluffy rice cake..more',
+        desc:
+            'A healthy breakfast item and an authentic south indian delicacy! Steamed and fluffy rice cake..more',
       ),
       RestaurantDetail(
         title: 'Sambar Idly (2Pcs)',
@@ -30,7 +29,8 @@ class RestaurantDetail {
         title: 'Ghee Pongal',
         image: 'assets/images/food3.jpg',
         price: 'Rs85',
-        desc: 'Cute, button idlis with authentic. South Indian sambar and coconut chutney gives the per..more',
+        desc:
+            'Cute, button idlis with authentic. South Indian sambar and coconut chutney gives the per..more',
       ),
       RestaurantDetail(
         title: 'Boori (1Set)',
@@ -46,7 +46,8 @@ class RestaurantDetail {
         title: 'Mini Idly with Sambar',
         image: 'assets/images/food6.jpg',
         price: 'Rs85',
-        desc: 'Cute, button idlis with authentic. South Indian sambar and coconut chutney gives the per..more',
+        desc:
+            'Cute, button idlis with authentic. South Indian sambar and coconut chutney gives the per..more',
       ),
     ];
   }
@@ -56,7 +57,8 @@ class RestaurantDetail {
       RestaurantDetail(
         title: 'Plain Dosa',
         price: 'Rs30',
-        desc: 'A healthy breakfast item and an authentic south indian delicacy!',
+        desc:
+            'A healthy breakfast item and an authentic south indian delicacy!',
       ),
       RestaurantDetail(
         title: 'Rava Dosa',
@@ -65,7 +67,8 @@ class RestaurantDetail {
       RestaurantDetail(
         title: 'Onion Dosa',
         price: 'Rs85',
-        desc: 'Cute, button dosas with authentic. South Indian sambar and coconut chutney gives the per..more',
+        desc:
+            'Cute, button dosas with authentic. South Indian sambar and coconut chutney gives the per..more',
       ),
       RestaurantDetail(
         title: 'Onion Uttapam',
