@@ -4,8 +4,8 @@ import 'package:swiggy_ui/utils/ui_helper.dart';
 
 class SearchFoodListItemView extends StatelessWidget {
   const SearchFoodListItemView({
-    Key key,
-    @required this.food,
+    Key? key,
+    required this.food,
   }) : super(key: key);
 
   final SpotlightBestTopFood food;
@@ -43,10 +43,10 @@ class SearchFoodListItemView extends StatelessWidget {
               children: <Widget>[
                 Text(
                   food.name,
-                  style: Theme.of(context).textTheme.subtitle2.copyWith(fontSize: 15.0),
+                  style: Theme.of(context).textTheme.subtitle2!.copyWith(fontSize: 15.0),
                 ),
                 Text(food.desc,
-                    style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey[600], fontSize: 13.5)),
+                    style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.grey[600], fontSize: 13.5)),
                 UIHelper.verticalSpaceSmall(),
                 Row(
                   children: <Widget>[
