@@ -54,14 +54,14 @@ class InTheSpotlightView extends StatelessWidget {
                 UIHelper.horizontalSpaceSmall(),
                 Text(
                   'In the Spotlight!',
-                  style: Theme.of(context).textTheme.headline4.copyWith(fontSize: 20.0),
+                  style: Theme.of(context).textTheme.headline4!.copyWith(fontSize: 20.0),
                 )
               ],
             ),
             UIHelper.verticalSpaceExtraSmall(),
             Text(
               'Explore sponsored partner brands',
-              style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey),
+              style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.grey),
             ),
           ],
         ),
