@@ -1,5 +1,5 @@
+import 'package:card_swiper/card_swiper.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_swiper/flutter_swiper.dart';
 import 'package:swiggy_ui/widgets/responsive.dart';
 
 import '../groceries/grocery_screen.dart';
