@@ -25,14 +25,14 @@ class TopOffersViews extends StatelessWidget {
                     UIHelper.horizontalSpaceExtraSmall(),
                     Text(
                       'Top Offers',
-                      style: Theme.of(context).textTheme.headline4.copyWith(fontSize: 20.0),
+                      style: Theme.of(context).textTheme.headline4!.copyWith(fontSize: 20.0),
                     ),
                     Spacer(),
                     Row(
                       children: <Widget>[
                         Text(
                           'SEE ALL',
-                          style: Theme.of(context).textTheme.bodyText1.copyWith(fontWeight: FontWeight.bold),
+                          style: Theme.of(context).textTheme.bodyText1!.copyWith(fontWeight: FontWeight.bold),
                         ),
                         UIHelper.horizontalSpaceExtraSmall(),
                         ClipOval(
@@ -55,7 +55,7 @@ class TopOffersViews extends StatelessWidget {
                 UIHelper.verticalSpaceExtraSmall(),
                 Text(
                   'Get 20-50% Off',
-                  style: Theme.of(context).textTheme.bodyText1.copyWith(color: Colors.grey),
+                  style: Theme.of(context).textTheme.bodyText1!.copyWith(color: Colors.grey),
                 ),
               ],
             ),
