@@ -8,7 +8,7 @@ class VegBadgeView extends StatelessWidget {
       height: 15.0,
       width: 15.0,
       decoration: BoxDecoration(
-        border: Border.all(color: Colors.green[800]),
+        border: Border.all(color: Colors.green[800]!),
       ),
       child: ClipOval(
         child: Container(
